Damage Implementation (UNFINISHED)
Created a Meta Attribute: IncomingDamage
On the AttributeSet: Reduced Health by IncomingDamage

To make Damage dependent on the Spell cast:
Changed GE_Damage to Set by Caller, and added a Damage GameplayTag, Assigned SetByCallerMagnitude in AuraProjectileSpell, added the Damage Tag to the GE, added Damage Scalable Float onto AuraGameplayAbility (All Abilities now have a Damage property), created a Curve Table for Damage, with a Curve for Firebolt.
Therefore, Each Ability's Damage is scaled by a curve in C++ on the Attribute Set, which is being called when IncomingDamage is changed

For Hit Reaction Gameplay Effect:
Created GE_HitReact, Created a native GameplayTag Effects.HitReact, Added Effects.HitReact as a GrantedTag onto GE_HitReact. Registered a callback onto the Enemy's ASC to listen for Effects.HitReact to be added or removed.

For Hit Reaction Gameplay Ability (Which Applies the GE):
Created GA_HitReact, Applied GE_HitReact, Play the Actor's corresponding HitReactMontage (which was added as a variable onto AuraCharacterBase and accessed through the CombatInterface).

To Activate the Hit React Ability:
Added GiveStartupAbilities onto the AuraAbilitySystemLibrary. In it, access the GameMode, and then through the GameMode, access CharacterClassInfo, the Data Asset that holds the array of GAs to grant.
If the Incoming Damage is fatal:
call Die, and then MulticastDie to replicate onto clients
If !fatal:
Try to activate GA_HitReact through the Effects.HitReact GameplayTag (which is added to GA_HitReact)

Implementing Die:
Weapon and Mesh ragdoll, then Dissolve() is called, which applies a dynamic material instance over a timeline implemented in blueprint to dissolve the Weapon and the Mesh

Created a floating text widget and DamageTextComponent (Widget Component) with animation to reflect Incoming damage on enemies (TODO: Implementation in progress)